requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
markitdown[all]==0.1.2
mangum==0.17.0
pydantic==2.5.0
httpx==0.25.2
python-magic==0.4.27
psutil==5.9.6
openai>=1.0.0